+++ src/missions/bring_product.lua
@@ -47,7 +47,7 @@ Missions.bringProduct = function(self, station, config)
             local sellProductMenu
             sellProductMenu = function(amount)
                 amount = amount or 0
-                return function(self, comms_target, comms_source)
+                return function(comms_target, comms_source)
                     if amount > 0 then
                         if not Player:hasStorage(comms_source) then
                             logWarning("Player Ship has no storage, but an action to sell " .. product:getName() .. " was called. Probably an issue in your comms script.")